ismail/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
mysqlclient
python-dotenv
PyDrive2
google-api-python-client
google-auth
Werkzeug 